requirements-dev.txt
pytest-homeassistant-custom-component
requests-mock
pytest-cov
